requirements.txt
alembic==0.9.8
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
flask-paginate==0.5.1
Flask-Script==2.0.6

SQLAlchemy==1.2.3
click==6.7
dominate==2.3.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
python-dateutil==2.7.2
python-editor==1.0.3
six==1.11.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
requests==2.18.4
certifi==2018.1.18
chardet==3.0.4
idna==2.6
urllib3==1.22
psycopg2-binary==2.7.4
ftfy==5.3.0
requests-file==1.4.3
tldextract==2.2.0
wcwidth==0.1.7
EasyProcess==0.2.3
PyVirtualDisplay==0.2.1
selenium==3.11.0
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
kombu==4.1.0
pytz==2018.4
redis==2.10.6
vine==1.1.4
eventlet==0.22.1
greenlet==0.4.13
